명시적인 auto boxing과 auto unboxing으로 변경

+++ gradle/src/main/java/com/naver/cafe/javachobostudy/journy002/AutoBoxingUnboxing.java
@@ -2,8 +2,8 @@ package com.naver.cafe.javachobostudy.journy002;
 
 public class AutoBoxingUnboxing {
     public static void main(String[] args) {
-        Integer integer = 12345;    // auto boxing
-        int i = integer;            // auto unboxing
+        Integer integer = Integer.valueOf(12345);
+        int i = integer.intValue();
 
         System.out.println(integer);
         System.out.println(i);
